add templates file, fix compile
